+++ lib/src/loaders/assets.dart
@@ -4,11 +4,8 @@ import '../../nations.dart';
 import 'base.dart';
 
 /// * to load lang files from 'assets/lang/'
-class NationsJsonLoader extends NationsLoader {
-  const NationsJsonLoader();
+class NationsAssetsLoader extends NationsLoader {
+  const NationsAssetsLoader();
   @override
   Future<Map<String, dynamic>?> load(locale) => loadLocaleTranslation(locale);
-
-  @override
-  String toString() => 'NationsJsonLoader';
 }