+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.0.3
 
+- ⚠ Breaking :: rename `NationsJsonLoader` to `NationsAssetsLoader`
+
 ## 0.0.2
 
 - ⚠ Breaking :: remove translation methods , we will stick with extension methods to unify the package usage in all projects